# Conflicts:
#	test/optimization/test_converters.py

+++ qiskit/optimization/converters/integer_to_binary_converter.py
@@ -81,16 +81,18 @@ class IntegerToBinaryConverter:
         upper_bounds = self._src.variables.get_upper_bounds()
         for i, variable in enumerate(names):
             typ = types[i]
-            if typ == 'I':
-                new_vars: List[Tuple[str, int]] = self._encode_var(name=variable,
-                                                                   lower_bound=lower_bounds[i],
-                                                                   upper_bound=upper_bounds[i])
+            if typ == "I":
+                new_vars: List[Tuple[str, int]] = self._encode_var(
+                    name=variable, lower_bound=lower_bounds[i], upper_bound=upper_bounds[i]
+                )
                 self._conv[variable] = new_vars
-                self._dst.variables.add(names=[new_name for new_name, _ in new_vars],
-                                        types='B' * len(new_vars))
+                self._dst.variables.add(
+                    names=[new_name for new_name, _ in new_vars], types="B" * len(new_vars)
+                )
             else:
-                self._dst.variables.add(names=[variable], types=typ,
-                                        lb=[lower_bounds[i]], ub=[upper_bounds[i]])
+                self._dst.variables.add(
+                    names=[variable], types=typ, lb=[lower_bounds[i]], ub=[upper_bounds[i]]
+                )
 
         self._substitute_int_var()
 
@@ -194,8 +196,9 @@ class IntegerToBinaryConverter:
 
             lin_expr.append(sparse_pair)
 
-        self._dst.linear_constraints.add(lin_expr, linear_sense, linear_rhs, linear_ranges,
-                                         linear_names)
+        self._dst.linear_constraints.add(
+            lin_expr, linear_sense, linear_rhs, linear_ranges, linear_names
+        )
 
         # TODO: add quadratic constraints
         if self._src.quadratic_constraints.get_num() > 0:
@@ -218,7 +221,7 @@ class IntegerToBinaryConverter:
 
     def _decode_var(self, names, vals) -> List[int]:
         # decode integer values
-        sol = {name: int(vals[i]) for i, name in enumerate(names)}
+        sol = {name: float(vals[i]) for i, name in enumerate(names)}
         new_vals = []
         for name in self._src.variables.get_names():
             if name in self._conv:
